+++ WebContent/find_idpw.jsp
@@ -57,8 +57,8 @@
                         <label class="account-id-error" for="account-id" style="font-size: 14px; font-weight: 400; line-height: 20px; top: -20px; width: 100%; position: absolute; margin: 0; left:0;  color: #757575;"></label>
                     </div>
                     <div class="form-group">
-                        <input class="form-control" id="account-email" type="email" placeholder="예) name@example.com">
-                        <label class="account-email-error" for="account-email" style="font-size: 14px; font-weight: 400; line-height: 20px; top: -20px; width: 100%; position: absolute; margin: 0; left:0;  color: #757575;">이메일</label>
+                        <input class="form-control" id="account-email-pw" type="email" placeholder="예) name@example.com">
+                        <label class="account-email-pw-error" for="account-email" style="font-size: 14px; font-weight: 400; line-height: 20px; top: -20px; width: 100%; position: absolute; margin: 0; left:0;  color: #757575;">이메일</label>
                     </div>
                 </div>
                 <div class="input-info-footer">
@@ -67,7 +67,7 @@
             <div class="row-footer">
                 <div class="row">
                 <!--회사소개  -->
-                    <a style="font-size: 5px;" href="http://lawcompany.co.kr/">
+                    <a style="font-size: 5px;" href="http://localhost:9000/kovengerss/compani.jsp">
                         (C) Law&Company Co., Ltd.
                     </a>
                 </div>
@@ -78,38 +78,68 @@
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 <script>
 
-     let emailCheck = $(".account-email-error");
-	let $email = $("#account-email");
-	
-	
-	$email.on("keyup", function(){
-		if($email.val() == ""){
-			emailCheck.css("color","red");
-			emailCheck.text("이메일은 반드시 입력해야합니다.");
-		}
-		
-		if($email.val() !== ""){
-			emailCheck.css("color","rgb(99 193 76)");
-			emailCheck.text("이메일 입력완료.");
-		}
-	});
-	
-	
-    let idCheck = $(".account-id-error");
-	let $id = $("#account-id");
-	
-	
-	$id.on("keyup", function(){
-		
-		if($id.val() == ""){
-			idCheck.css("color","red");
-			idCheck.text("아이디는 반드시 입력해야합니다.");
-		}
-		
-		if($id.val() !== ""){
-			idCheck.css("color","rgb(99 193 76)");
-			idCheck.text("아이디 입력완료.");
-		}
-	});
+var userIdCheck = RegExp(/^[A-Za-z0-9_\-]{5,20}$/);
+var nameCheck = RegExp(/^[가-힣]{2,6}$/);
+var emailCheck2 = RegExp(/^[A-Za-z0-9_\.\-]+@[A-Za-z0-9\-]+\.[A-Za-z0-9\-]+/);
+var emailCheckP = RegExp(/^[A-Za-z0-9_\.\-]+@[A-Za-z0-9\-]+\.[A-Za-z0-9\-]+/);
+/* 이메일 유효성 검사 */    	
+let emailCheck = $(".account-email-error");
+let $email = $("#account-email");
+$email.on("keyup", function(){
+	if(emailCheck2.test($email.val())){
+		emailCheck.css("color","rgb(99 193 76)");
+		emailCheck.text("이메일 입력완료.");
+	} else if(nameCheck.test($email.val())){
+		emailCheck.css("color","red");
+		emailCheck.text("한글은 입력 할 수 없습니다.");
+	} else if($email.val() ==""){
+		emailCheck.css("color","red");
+		emailCheck.text("이메일을 입력해주세요");
+	}
+	else{
+		emailCheck.css("color","red");
+		emailCheck.text("이메일 형식에 맞지 않습니다.");
+	}
+});
+
+let emailCheckpw = $(".account-email-pw-error");
+let $emailPw = $("#account-email-pw");
+$emailPw.on("keyup", function(){
+	if(emailCheckP.test($emailPw.val())){
+		emailCheckpw.css("color","rgb(99 193 76)");
+		emailCheckpw.text("이메일 입력완료.");
+	} else if(nameCheck.test($emailPw.val())){
+		emailCheckpw.css("color","red");
+		emailCheckpw.text("한글은 입력 할 수 없습니다.");
+	} else if($emailPw.val() ==""){
+		emailCheckpw.css("color","red");
+		emailCheckpw.text("이메일을 입력해주세요");
+	}
+	else{
+		emailCheckpw.css("color","red");
+		emailCheckpw.text("이메일 형식에 맞지 않습니다.");
+	}
+});
+
+
+
+/* 아이디 유효성 검사  */    	
+let idCheck = $(".account-id-error");
+let $id = $("#account-id");
+$id.on("keyup", function(){
+	if(userIdCheck.test($id.val())){
+		idCheck.css("color","rgb(99 193 76)");
+		idCheck.text("아이디 입력완료.");
+	}else if($id.val() == ""){
+    	idCheck.css("color","red");
+    	idCheck.text("아이디는 반드시 입력해야합니다.");
+	}
+	else{
+		idCheck.css("color","red");
+		idCheck.text("아이디 형식에 맞지 않습니다.");
+	}
+});
+
+
 </script>
 </html>
