+++ src/com/lawknow/domain/dao/UserReviewDAO.java
@@ -1,4 +1,5 @@
-package com.userwrite.domain.dao;
+package com.lawknow.domain.dao;
+
 
 import java.sql.Date;
 
@@ -6,28 +7,28 @@ import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 
 import com.mybatis.config.MyBatisConfig;
-import com.userwrite.domain.vo.UserWriteVO;
+import com.userreview.domain.vo.UserReviewVO;
 
-public class UserWriteDAO {
+public class UserReviewDAO {
 	SqlSessionFactory sqlSessionFactory = MyBatisConfig.getSqlSessionFactory();
 	SqlSession sqlSession;
 	
-	public UserWriteDAO() {
+	public UserReviewDAO() {
 		sqlSession = sqlSessionFactory.openSession(true);
 	}
 	
+	//리뷰 번호
+	public int reviewNum(UserReviewVO userReviewVO) {
+		return 0;
+	}
+	
 	//사건 번호
-	public int requestNum(UserWriteVO userVO) {
+	public int requestNum(UserReviewVO userVO) {
 		return 0;
 	}
 	
-	//의뢰인 글 번호
-	public int userWirteNum(UserWriteVO userVO) {
-		return 0;
-	}
-	
-	//의뢰인 글 날짜
-	public Date userWriteDate(UserWriteVO userVO) {
+	//리뷰 날짜
+	public Date reviewWriteDate(UserReviewVO userVO) {
 		return null;
 	}
 }
