+++ src/com/lawknow/domain/dao/UserWriteDAO.java
@@ -1,5 +1,4 @@
-package com.userreview.domain.dao;
-
+package com.lawknow.domain.dao;
 
 import java.sql.Date;
 
@@ -7,28 +6,28 @@ import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 
 import com.mybatis.config.MyBatisConfig;
-import com.userreview.domain.vo.UserReviewVO;
+import com.userwrite.domain.vo.UserWriteVO;
 
-public class UserReviewDAO {
+public class UserWriteDAO {
 	SqlSessionFactory sqlSessionFactory = MyBatisConfig.getSqlSessionFactory();
 	SqlSession sqlSession;
 	
-	public UserReviewDAO() {
+	public UserWriteDAO() {
 		sqlSession = sqlSessionFactory.openSession(true);
 	}
 	
-	//리뷰 번호
-	public int reviewNum(UserReviewVO userReviewVO) {
-		return 0;
-	}
-	
 	//사건 번호
-	public int requestNum(UserReviewVO userVO) {
+	public int requestNum(UserWriteVO userVO) {
 		return 0;
 	}
 	
-	//리뷰 날짜
-	public Date reviewWriteDate(UserReviewVO userVO) {
+	//의뢰인 글 번호
+	public int userWirteNum(UserWriteVO userVO) {
+		return 0;
+	}
+	
+	//의뢰인 글 날짜
+	public Date userWriteDate(UserWriteVO userVO) {
 		return null;
 	}
 }
